Fixed issues with conda in pyenv activation
For some version of conda, the command is not available even if we
active the required pyenv virtualenv. Now, all the miniconda virtualenv
are managed by conda and the rest of the virtualenv is managed by pyenv
as before. This however add extra complexity is python management.
Hopefully I will solve it by redefining `pa` and `pd` commands.

TODO: redefine `pa` and `pd` command to activate and deactivate the
virtualenv transparently for pyenv and conda virtualenvs.